Chore/slice workflow (#18351)
