+++ web/app/components/workflow/context.tsx
@@ -2,19 +2,24 @@ import {
   createContext,
   useRef,
 } from 'react'
-import { createWorkflowStore } from './store'
+import {
+  createWorkflowStore,
+} from './store'
+import type { StateCreator } from 'zustand'
+import type { WorkflowSliceShape } from '@/app/components/workflow-app/store/workflow/workflow-slice'
 
 type WorkflowStore = ReturnType<typeof createWorkflowStore>
 export const WorkflowContext = createContext<WorkflowStore | null>(null)
 
-type WorkflowProviderProps = {
+export type WorkflowProviderProps = {
   children: React.ReactNode
+  injectWorkflowStoreSliceFn?: StateCreator<WorkflowSliceShape>
 }
-export const WorkflowContextProvider = ({ children }: WorkflowProviderProps) => {
+export const WorkflowContextProvider = ({ children, injectWorkflowStoreSliceFn }: WorkflowProviderProps) => {
   const storeRef = useRef<WorkflowStore | undefined>(undefined)
 
   if (!storeRef.current)
-    storeRef.current = createWorkflowStore()
+    storeRef.current = createWorkflowStore({ injectWorkflowStoreSliceFn })
 
   return (
     <WorkflowContext.Provider value={storeRef.current}>
