+++ web/app/components/workflow/hooks/index.ts
@@ -5,7 +5,6 @@ export * from './use-nodes-data'
 export * from './use-nodes-sync-draft'
 export * from './use-workflow'
 export * from './use-workflow-run'
-export * from './use-workflow-template'
 export * from './use-checklist'
 export * from './use-selection-interactions'
 export * from './use-panel-interactions'
@@ -16,3 +15,4 @@ export * from './use-workflow-variables'
 export * from './use-shortcuts'
 export * from './use-workflow-interactions'
 export * from './use-workflow-mode'
+export * from './use-format-time-from-now'
